requirements.txt
Package                Version
---------------------- -----------
Jinja2                 3.1.4
praw                   7.7.1
python-dotenv          1.0.1
python3-discogs-client 2.7


